+++ packages/platform/tests/main.ts
@@ -1,5 +1,6 @@
 import Platform from '../src/index'
 import { expect } from 'chai'
+import { createCallerFactory } from '../src/trpc.js'
 
 import { version } from '../package.json'
 
@@ -11,17 +12,17 @@ describe('Platform', () => {
             expect(platform).to.be.an.instanceOf(Platform)
         })
         
-        // it('should be able to query platform version', async function(){
-        //     const platform = new Platform()
-        //     const caller = platform.appRouter.createCaller({})
+        it('should be able to query platform version', async function(){
+            const platform = new Platform()
+            const caller = createCallerFactory(platform.appRouter)({})
 
-        //     const response = await caller.version()
-        //     expect(response).to.equal(version)
-        // })
+            const response = await caller.version()
+            expect(response).to.equal(version)
+        })
 
         it('should be able to ping appRouter', async function(){
             const platform = new Platform()
-            const caller = platform.appRouter.createCaller({})
+            const caller = createCallerFactory(platform.appRouter)({})
 
             const response = await caller.ping()